feat: clean-up
- add set reward distributor with new entry point call data
- clean up user flow for both set distributor and add reward
